<FIX> view 관련 역할 분리, random value역할 분리
