+++ src/main/java/domain/Rule.java
@@ -1,0 +1,5 @@
+package domain;
+
+public interface Rule {
+    public boolean canGo();
+}
